requirements.txt
absl-py==1.0.0
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
click==8.0.1
colorama==0.4.4
cycler==0.11.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.9
django-extensions==3.1.5
entrypoints==0.3
Faker==8.10.3
Flask==2.0.1
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.2
importlib-metadata==4.8.2
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.6.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pyupbit==0.2.19
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
requests-oauthlib==1.3.0
riotwatcher==3.1.2
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.7.1
setuptools-scm==6.3.2
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.0.5
typing-extensions==4.0.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.1
wrapt==1.13.3
yapf==0.31.0
zipp==3.6.0
